#22 MatchingResponseDto 거리 타입 int로 변경

+++ src/main/java/clonecoding/tinder/matching/model/dto/MatchingResponseDto.java
@@ -14,7 +14,7 @@ public class MatchingResponseDto {
     private Long memberId;
     private String nickName;
     private String profile;
-    private double distance;
+    private int distance;
     private int age;
     private Long roomId;
 }
